Extract message factory
